requirements.txt
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
pytesseract==0.3.7
tqdm==4.62.3
transformers==4.11.2